Rename: src/components/Post 경로를 src/components/PostSection으로 변경
