+++ src/components/PostSection/CategoryInput.tsx
