+++ src/components/PostSection/InputWithEnter.tsx
