+++ src/components/PostSection/TeamInput.tsx
